Remove usages of static_cast<enum> when outputing to stream.
Bug: 1318236
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4673703
Auto-Submit: Xiaoling Bao <xiaolingbao@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Xiaoling Bao <xiaolingbao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1167652}

+++ chrome/updater/policy/policy_fetcher.cc
@@ -24,6 +24,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "chrome/updater/device_management/dm_storage.h"
 #include "chrome/updater/policy/dm_policy_manager.h"
 #include "chrome/updater/policy/service.h"
+#include "chrome/updater/util/util.h"
 #include "third_party/abseil-cpp/absl/types/optional.h"
 #include "url/gurl.h"
 
@@ -141,7 +142,7 @@ PolicyFetcher::OnFetchPolicyRequestComplete(
               if (result != DMClient::RequestResult::kSuccess)
                 LOG(WARNING)
                     << "DMClient::ReportPolicyValidationErrors failed: "
-                    << static_cast<int>(result);
+                    << result;
             })));
   }
